% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtrim.R
\name{bib_add}
\alias{bib_add}
\title{Add to the project-specific bibtex file}
\usage{
bib_add(biblib_file, bibref_file, citations)
}
\arguments{
\item{biblib_file}{The location of the project-specific bibtex file.}

\item{bibref_file}{The location of the reference bibtex file.}

\item{citations}{The citations used in the text to add to the
project-specific bibtex file.}
}
\value{
Creates the revised project-specific bibtex file.
}
\description{
Adds the new references from a set of citations to a project-specific bibtex
file from a reference bibtex file.
}
\examples{
\dontrun{
bib_add("Introduction.Rmd")
}
}
