% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtrim.R
\name{bib_trim}
\alias{bib_trim}
\title{Trim the project-specific bibtex}
\usage{
bib_trim(biblib_file, citations)
}
\arguments{
\item{biblib_file}{The location of the project-specific bibtex file.}

\item{citationsThe}{citations used in the text to check against
the project-specific bibtex file.}
}
\value{
Creates the revised project-specific bibtex file.
}
\description{
Trims the project-specific bibtex file so that it includes only a given set of references
}
