% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtrim.R
\name{bibref_check}
\alias{bibref_check}
\title{Check which reference bibtex file exists}
\usage{
bibref_check(bibref_file_locations)
}
\arguments{
\item{bibref_file_locations}{A vector of potential file locations for the
reference bibtex file}
}
\value{
The name of the reference bibtex file which exists
}
\description{
This function is useful for people like me who write across different
computers, upon which the bibtex file is saved in different locations (such
as automatically by Mendeley). This function searches for which of the given
locations contains the reference file.
}
