% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtrim.R
\name{cites_get}
\alias{cites_get}
\title{Get Citations}
\usage{
cites_get(rmdfile)
}
\arguments{
\item{rmdfile}{The name of the file}
}
\value{
A vector of all the bibtex keys from the Rmd file.
}
\description{
Get the citations from an Rmd file.
}
\examples{
\dontrun{
citations <- cites_get("Introduction.Rmd")
}

}
