% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtrim.R
\name{cites_get_all}
\alias{cites_get_all}
\title{Get citations from multiple Rmd files}
\usage{
cites_get_all(rmdfiles)
}
\arguments{
\item{rmdfiles}{A vector of Rmd files.}
}
\value{
A vector of all unique bibtex keys used across all the Rmd files.
}
\description{
Extracts all citations from several Rmd files
}
