% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtrim.R
\name{rmd_get_all}
\alias{rmd_get_all}
\title{Find all Rmd files}
\usage{
rmd_get_all(rmdfolder = getwd(), recursive = T, ...)
}
\arguments{
\item{recursive}{Logical - should the search be recursive?}

\item{...}{Extra arguments for \code{list.files}.}

\item{folder}{The folder to search within}
}
\value{
A vector of all the Rmd files in a given folder
}
\description{
Finds all Rmd files in a given folder
}
